Built a dice rolling circuit that plays music
First project was about getting acquainted working with servo motors and using a potentiometer to control its angle.

The second project was about making a program that rolled a dice and displayed the result using a servo. In addition, sounds were added using a piezosumrer which in the end was able to play different melodies.